% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_loading.R, R/load_values_dictionary.R
\name{load_values_dictionary}
\alias{load_values_dictionary}
\title{Load Values Dictionary}
\usage{
load_values_dictionary(dictionary_path = NULL)

load_values_dictionary(dictionary_path = NULL)
}
\arguments{
\item{dictionary_path}{Optional. The path to the values dictionary CSV file. If not provided, the default path will be used.}
}
\value{
A data frame with the values dictionary.

A data frame with the values dictionary.
}
\description{
This function loads the values dictionary from the package's extdata directory.

This function loads the values dictionary from the package's extdata directory.
}
